requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.9
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
alembic==1.16.5
python-dotenv==1.1.1